pyproject.toml
[project]
name = "stock-ml"
version = "0.1.0"
description = "Making Stock investing ML"
authors = [
    {name = "Juwon Lee",email = "juwonlee3465@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "yfinance (>=0.2.66,<0.3.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "tensorflow (>=2.20.0,<3.0.0)",
    "textblob (>=0.19.0,<0.20.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "fredapi (>=0.5.2,<0.6.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "transformers (>=4.57.1,<5.0.0)",
    "torch (>=2.9.0,<3.0.0) ; python_version >= \"3.10\" and python_version < \"3.15\"",
    "tf-keras (>=2.20.1,<3.0.0)",
    "duckdb (>=1.4.1,<2.0.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "html5lib (>=1.1,<2.0)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "streamlit (>=1.51.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ipykernel (>=7.1.0,<8.0.0)"
]
